requirements.txt
asgiref==3.7.2
Django==4.2.7
django-filter==23.3
django-phonenumber-field==7.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
flake8==6.1.0
iniconfig==2.0.0
isort==5.12.0
mccabe==0.7.0
packaging==23.2
phonenumbers==8.13.25
pluggy==1.3.0
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pyflakes==3.1.0
PyJWT==2.8.0
pytest==7.4.3
pytest-django==4.7.0
pytz==2023.3.post1
sqlparse==0.4.4
